pyproject.toml
[project]
name = "code_chat_bot"
version = "0.1.0"
description = "Multiple AI ChatBot - OpenAI, MistralAI, Anthropic and Cohere"
readme = "README.md"
authors = [
    {name = "CodeHalwell", email = "your-email@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    # Core AI and ML
    "accelerate>=0.30.0",
    "anthropic>=0.30.0",
    "huggingface-hub>=0.23.0",
    "numpy>=1.26.4",
    "transformers>=4.41.0",

    # LangChain ecosystem
    "langchain>=0.3.27",
    "langchain-anthropic>=1.1.0",
    "langchain-community>=0.0.32",
    "langchain-core>=0.2.0",
    "langchain-google-vertexai>=1.0.2",
    "langchain-mistralai>=0.1.3",
    "langchain-openai>=0.1.2",
    "langchain-text-splitters>=0.0.2",
    "langchain-chroma>=0.1.1",
    "langchain-google-genai>=1.0.6",
    "langsmith>=0.2.0",

    # AI Provider SDKs
    "mistralai>=0.2.0",
    "openai>=1.30.0",
    "cohere>=1.0.0",
    "google-generativeai>=0.5.4",

    # Web framework
    "streamlit>=1.34.0",

    # Data processing
    "pandas>=2.2.2",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.2",
    "tiktoken>=0.7.0",

    # Database support
    "pymongo>=4.7.2",
    "psycopg2-binary>=2.9.9",
    "firebase-admin>=6.5.0",

    # Voice capabilities
    "SpeechRecognition>=3.10.4",
    "pyttsx3>=2.90",
    "gTTS>=2.5.1",

    # NLP and analysis
    "spacy>=3.7.4",
    "textblob>=0.18.0",
    "langdetect>=1.0.9",

    # Monitoring and logging
    "loguru>=0.7.2",
    "prometheus-client>=0.20.0",

    # Additional utilities
    "pydantic>=2.7.1",
    "requests>=2.32.2",
    "beautifulsoup4>=4.12.3",
    "chromadb>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88